create common folder for common funcs
